requirements.txt
matplotlib
GitPython==2.1.8
pandas==0.22.0
scikit-learn==0.19.1
numpy==1.14.0
plotly==2.5.1
scipy==1.0.0